requirements.txt
django==3.0.8
pillow==8.3.1
psycopg2-binary==2.9.1
django-environ==0.4.5
django-crispy-forms==1.9.1
django-allauth==0.43.0
djangorestframework
markdown 
django-filter
